Replace MonthLoader with EventsLoader, adapt class names
